move eventTotals to own component
